06_sql_nosql/requirements.txt
flask 
mysql-connector-python 
pymongo